State pattern refactoring
